requirements.txt
os
glob
Flask
gunicorn
joblib
scikit-learn
python_speech_features
pydub
scipy.io.wavfile
werkzeug.utils
ffmpeg